django/requirements.txt
bcrypt==3.2.0
Django==3.1.6
djangorestframework==3.12.2
requests==2.25.1